feat: database improvements - refresh db on startup if shorter - disable dozer - update application properties

+++ src/main/java/com/github/solairerove/harald/domain/repository/PostRepository.java
@@ -3,6 +3,6 @@ package com.github.solairerove.harald.domain.repository;
 import com.github.solairerove.harald.domain.model.Post;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostRepository extends CrudRepository<Post, Integer> {
+public interface PostRepository extends CrudRepository<Post, Long> {
 
 }
